plat-sam: matrix: update code to be reuseable for sama7g5
Besides sama5d2, sama7g5 also has the matrix. Following changes are done
to make the code reuseable for supporting sama7g5:
- move definition of "peri_security_array[]" from matrix.c to main.c
- replace "matrix32_base()" and "matrix64_base()" with "matrix_base()"
- update code according to the above changes

Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
