+++ core/arch/arm/plat-sam/matrix.c
@@ -32,468 +32,15 @@
 #include <kernel/pm.h>
 #include <kernel/panic.h>
 #include <matrix.h>
+#include <mm/core_memprot.h>
+#include <mm/core_mmu.h>
 #include <platform_config.h>
-#include <sama5d2.h>
 #include <stdint.h>
 #include <tz_matrix.h>
 #include <trace.h>
 
-#define MATRIX_AXIMX	1
-#define MATRIX_H64MX	2
-#define MATRIX_H32MX	3
-
-#define SECURITY_TYPE_AS	1
-#define SECURITY_TYPE_NS	2
-#define SECURITY_TYPE_PS	3
-
 #define WORLD_NON_SECURE	0
 #define WORLD_SECURE		1
-
-#define MATRIX_SPSELR_COUNT	3
-#define MATRIX_SLAVE_COUNT	15
-
-struct peri_security {
-	unsigned int peri_id;
-	unsigned int matrix;
-	unsigned int security_type;
-	paddr_t addr;
-};
-
-static const struct peri_security peri_security_array[] = {
-	{
-		.peri_id = AT91C_ID_PMC,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_PMC,
-	},
-	{
-		.peri_id = AT91C_ID_ARM,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_PIT,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_PITC,
-	},
-	{
-		.peri_id = AT91C_ID_WDT,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_WDT,
-	},
-	{
-		.peri_id = AT91C_ID_GMAC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_GMAC,
-	},
-	{
-		.peri_id = AT91C_ID_XDMAC0,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_XDMAC0,
-	},
-	{
-		.peri_id = AT91C_ID_XDMAC1,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_XDMAC1,
-	},
-	{
-		.peri_id = AT91C_ID_ICM,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_ICM,
-	},
-	{
-		.peri_id = AT91C_ID_AES,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_AES,
-	},
-	{
-		.peri_id = AT91C_ID_AESB,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_AESB,
-	},
-	{
-		.peri_id = AT91C_ID_TDES,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_TDES,
-	},
-	{
-		.peri_id = AT91C_ID_SHA,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SHA,
-	},
-	{
-		.peri_id = AT91C_ID_MPDDRC,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_MPDDRC,
-	},
-	{
-		.peri_id = AT91C_ID_MATRIX1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_MATRIX32,
-	},
-	{
-		.peri_id = AT91C_ID_MATRIX0,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_MATRIX64,
-	},
-	{
-		.peri_id = AT91C_ID_SECUMOD,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_SECUMOD,
-	},
-	{
-		.peri_id = AT91C_ID_HSMC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_HSMC,
-	},
-	{
-		.peri_id = AT91C_ID_PIOA,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_PIOA,
-	},
-	{
-		.peri_id = AT91C_ID_FLEXCOM0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_FLEXCOM0,
-	},
-	{
-		.peri_id = AT91C_ID_FLEXCOM1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_FLEXCOM1,
-	},
-	{
-		.peri_id = AT91C_ID_FLEXCOM2,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_FLEXCOM2,
-	},
-	{
-		.peri_id = AT91C_ID_FLEXCOM3,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_FLEXCOM3,
-	},
-	{
-		.peri_id = AT91C_ID_FLEXCOM4,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_FLEXCOM4,
-	},
-	{
-		.peri_id = AT91C_ID_UART0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_UART0,
-	},
-	{
-		.peri_id = AT91C_ID_UART1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_UART1,
-	},
-	{
-		.peri_id = AT91C_ID_UART2,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_UART2,
-	},
-	{
-		.peri_id = AT91C_ID_UART3,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_UART3,
-	},
-	{
-		.peri_id = AT91C_ID_UART4,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_UART4,
-	},
-	{
-		.peri_id = AT91C_ID_TWI0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_TWI0,
-	},
-	{
-		.peri_id = AT91C_ID_TWI1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_TWI1,
-	},
-	{
-		.peri_id = AT91C_ID_SDMMC0,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SDHC0,
-	},
-	{
-		.peri_id = AT91C_ID_SDMMC1,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SDHC1,
-	},
-	{
-		.peri_id = AT91C_ID_SPI0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SPI0,
-	},
-	{
-		.peri_id = AT91C_ID_SPI1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SPI1,
-	},
-	{
-		.peri_id = AT91C_ID_TC0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_TC0,
-	},
-	{
-		.peri_id = AT91C_ID_TC1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_TC1,
-	},
-	{
-		.peri_id = AT91C_ID_PWM,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_PWMC,
-	},
-	{
-		.peri_id = AT91C_ID_ADC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_ADC,
-	},
-	{
-		.peri_id = AT91C_ID_UHPHS,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_UDPHS,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_UDPHS,
-	},
-	{
-		.peri_id = AT91C_ID_SSC0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SSC0,
-	},
-	{
-		.peri_id = AT91C_ID_SSC1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SSC1,
-	},
-	{
-		.peri_id = AT91C_ID_LCDC,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_LCDC,
-	},
-	{
-		.peri_id = AT91C_ID_ISI,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_HXISI,
-	},
-	{
-		.peri_id = AT91C_ID_TRNG,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_TRNG,
-	},
-	{
-		.peri_id = AT91C_ID_PDMIC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_PDMIC,
-	},
-	{
-		.peri_id = AT91C_ID_IRQ,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_NS,
-	},
-	{
-		.peri_id = AT91C_ID_SFC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SFC,
-	},
-	{
-		.peri_id = AT91C_ID_SECURAM,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_SECURAM,
-	},
-	{
-		.peri_id = AT91C_ID_QSPI0,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_QSPI0,
-	},
-	{
-		.peri_id = AT91C_ID_QSPI1,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_QSPI1,
-	},
-	{
-		.peri_id = AT91C_ID_I2SC0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_I2SC0,
-	},
-	{
-		.peri_id = AT91C_ID_I2SC1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_I2SC1,
-	},
-	{
-		.peri_id = AT91C_ID_CAN0_INT0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_CAN1_INT0,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_CLASSD,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_CLASSD,
-	},
-	{
-		.peri_id = AT91C_ID_SFR,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SFR,
-	},
-	{
-		.peri_id = AT91C_ID_SAIC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_SAIC,
-	},
-	{
-		.peri_id = AT91C_ID_AIC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_NS,
-		.addr = AT91C_BASE_AIC,
-	},
-	{
-		.peri_id = AT91C_ID_L2CC,
-		.matrix = MATRIX_H64MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_L2CC,
-	},
-	{
-		.peri_id = AT91C_ID_CAN0_INT1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_CAN1_INT1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_GMAC_Q1,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_GMAC_Q2,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_PIOB,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_PIOB,
-	},
-	{
-		.peri_id = AT91C_ID_PIOC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_PIOC,
-	},
-	{
-		.peri_id = AT91C_ID_PIOD,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_AS,
-		.addr = AT91C_BASE_PIOD,
-	},
-	{
-		.peri_id = AT91C_ID_SDMMC0_TIMER,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_SDMMC1_TIMER,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-	},
-	{
-		.peri_id = AT91C_ID_SYS,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SYSC,
-	},
-	{
-		.peri_id = AT91C_ID_ACC,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_ACC,
-	},
-	{
-		.peri_id = AT91C_ID_RXLP,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_RXLP,
-	},
-	{
-		.peri_id = AT91C_ID_SFRBU,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_SFRBU,
-	},
-	{
-		.peri_id = AT91C_ID_CHIPID,
-		.matrix = MATRIX_H32MX,
-		.security_type = SECURITY_TYPE_PS,
-		.addr = AT91C_BASE_CHIPID,
-	},
-};
 
 static void matrix_write(unsigned int base,
 			 unsigned int offset,
@@ -505,6 +52,20 @@ static void matrix_write(unsigned int base,
 static unsigned int matrix_read(int base, unsigned int offset)
 {
 	return io_read32(offset + base);
+}
+
+vaddr_t matrix_base(unsigned int matrix)
+{
+	unsigned int i = 0;
+	struct matrix *pmatrix = NULL;
+
+	do {
+		pmatrix = matrix_get(i++);
+		if (!pmatrix)
+			panic("Invalid matrix");
+	} while (pmatrix->matrix != matrix);
+
+	return io_pa_or_va_secure(&pmatrix->p, CORE_MMU_PGDIR_SIZE);
 }
 
 void matrix_write_protect_enable(unsigned int matrix_base)
@@ -531,35 +92,27 @@ void matrix_configure_slave_security(unsigned int matrix_base,
 
 static const struct peri_security *get_peri_security(unsigned int peri_id)
 {
-	unsigned int i;
+	unsigned int i = 0;
+	struct peri_security *p = NULL;
 
-	for (i = 0; i < ARRAY_SIZE(peri_security_array); i++) {
-		if (peri_id == peri_security_array[i].peri_id)
-			return &peri_security_array[i];
-	}
+	do {
+		p = peri_security_get(i++);
+		if (p && peri_id == p->peri_id)
+			break;
+	} while (p);
 
-	return NULL;
+	return p;
 }
 
 static int matrix_set_periph_world(unsigned int matrix, unsigned int peri_id,
 				   unsigned int world)
 {
-	unsigned int base;
-	unsigned int spselr;
-	unsigned int idx;
-	unsigned int bit;
+	unsigned int spselr = 0;
+	unsigned int idx = peri_id / 32;
+	unsigned int bit = 0x01 << (peri_id % 32);
+	unsigned int base = matrix_base(matrix);
 
-	idx = peri_id / 32;
 	if (idx > 3)
-		return -1;
-
-	bit = (0x01 << (peri_id % 32));
-
-	if (matrix == MATRIX_H32MX)
-		base = matrix32_base();
-	else if (matrix == MATRIX_H64MX)
-		base = matrix64_base();
-	else
 		return -1;
 
 	spselr = matrix_read(base, MATRIX_SPSELR(idx));
@@ -576,17 +129,19 @@ TEE_Result matrix_dt_get_id(const void *fdt, int node, unsigned int *id)
 {
 	unsigned int i = 0;
 	paddr_t pbase = 0;
+	struct peri_security *p = NULL;
 
 	pbase = fdt_reg_base_address(fdt, node);
 	if (pbase == DT_INFO_INVALID_REG)
 		return TEE_ERROR_BAD_PARAMETERS;
 
-	for (i = 0; i < ARRAY_SIZE(peri_security_array); i++) {
-		if (peri_security_array[i].addr == pbase) {
-			*id = peri_security_array[i].peri_id;
+	do {
+		p = peri_security_get(i++);
+		if (p && p->addr == pbase) {
+			*id = p->peri_id;
 			return TEE_SUCCESS;
 		}
-	}
+	} while (p);
 
 	return TEE_ERROR_ITEM_NOT_FOUND;
 }
@@ -605,17 +160,16 @@ int matrix_configure_periph_secure(unsigned int peri_id)
 int matrix_configure_periph_non_secure(unsigned int *peri_id_array,
 				       unsigned int size)
 {
-	unsigned int i;
-	unsigned int *peri_id_p;
-	unsigned int matrix;
-	unsigned int peri_id;
-	const struct peri_security *peripheral_sec;
-	int ret;
+	unsigned int i = 0;
+	unsigned int *peri_id_p = peri_id_array;
+	unsigned int matrix = 0;
+	unsigned int peri_id = 0;
+	const struct peri_security *peripheral_sec = NULL;
+	int ret = 0;
 
 	if (!peri_id_array || !size)
 		return -1;
 
-	peri_id_p = peri_id_array;
 	for (i = 0; i < size; i++) {
 		peripheral_sec = get_peri_security(*peri_id_p);
 		if (!peripheral_sec)
@@ -638,17 +192,6 @@ int matrix_configure_periph_non_secure(unsigned int *peri_id_array,
 }
 
 #ifdef CFG_PM_ARM32
-struct matrix_state {
-	uint32_t spselr[MATRIX_SPSELR_COUNT];
-	uint32_t ssr[MATRIX_SLAVE_COUNT];
-	uint32_t srtsr[MATRIX_SLAVE_COUNT];
-	uint32_t sassr[MATRIX_SLAVE_COUNT];
-	uint32_t meier;
-	uint32_t meimr;
-};
-
-static struct matrix_state matrix32_state;
-static struct matrix_state matrix64_state;
 
 static void matrix_save_regs(vaddr_t base, struct matrix_state *state)
 {
@@ -669,8 +212,11 @@ static void matrix_save_regs(vaddr_t base, struct matrix_state *state)
 
 static void matrix_suspend(void)
 {
-	matrix_save_regs(matrix32_base(), &matrix32_state);
-	matrix_save_regs(matrix64_base(), &matrix64_state);
+	unsigned int i = 0;
+	struct matrix *pmatrix = NULL;
+
+	for (pmatrix = matrix_get(i++); pmatrix; pmatrix = matrix_get(i++))
+		matrix_save_regs(matrix_base(pmatrix->matrix), &pmatrix->state);
 }
 
 static void matrix_restore_regs(vaddr_t base, struct matrix_state *state)
@@ -694,8 +240,12 @@ static void matrix_restore_regs(vaddr_t base, struct matrix_state *state)
 
 static void matrix_resume(void)
 {
-	matrix_restore_regs(matrix32_base(), &matrix32_state);
-	matrix_restore_regs(matrix64_base(), &matrix64_state);
+	unsigned int i = 0;
+	struct matrix *pmatrix = NULL;
+
+	for (pmatrix = matrix_get(i++); pmatrix; pmatrix = matrix_get(i++))
+		matrix_restore_regs(matrix_base(pmatrix->matrix),
+				    &pmatrix->state);
 }
 
 static TEE_Result matrix_pm(enum pm_op op, uint32_t pm_hint __unused,
